added files for form submission, added HTML
