+++ applicant.html
@@ -1,0 +1,38 @@
+<!DOCTYPE html>
+<html lang="en">
+<head>
+    <meta charset="UTF-8">
+    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+    <meta http-equiv="X-UA-Compatible" content="ie=edge">
+    <title>Applicant Detail</title>
+    <link rel="stylesheet" href="./src/main.css">
+    <link rel="stylesheet" href="./src/applicant.css">
+
+</head>
+<body>
+    <h1>Applicant Info</h1> 
+    <main>
+        <dl>
+            <dt>Name</dt>
+            <dd id="name"></dd>
+
+            <dt>Income</dt>
+            <dd id="income"></dd>
+
+            <dt>Burrito</dt>
+            <dd id="burrito"></dd>
+
+            <dt>Cat</dt>
+            <dd id="cat"></dd>
+
+            <dt>Cat Number</dt>
+            <dd id="cat-number"></dd>
+
+            <dt>Something Interesting</dt>
+            <dd id="something-interesting"></dd>
+        </dl>
+  </main>
+
+  <script type="module" src="./src/applicant.js"></script>
+</body>
+</html>
